Move legacy tts
